moved function class to submodule functions
