+++ src/f3dasm/__init__.py
@@ -18,11 +18,11 @@ from . import (data, functions, machinelearning, optimization, sampling,
                simulation)
 from ._show_versions import show_versions
 from .base.evaluator import Evaluator
-from .base.function import *
 from .base.utils import *
 from .design.design import *
 from .design.experimentdata import *
 from .design.parameter import *
+from .functions.function import *
 from .optimization.optimizer import *
 from .run_optimization import *
 from .sampling.sampler import *
