+++ tests/optimization/test_run_optimization.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from f3dasm.design.experimentdata import ExperimentData
-from f3dasm.base.function import Function
+from f3dasm.functions.function import Function
 from f3dasm.base.utils import make_nd_continuous_design
 from f3dasm.functions import (FUNCTIONS, FUNCTIONS_2D, FUNCTIONS_7D, Ackley,
                               Griewank, Levy, Rastrigin, Schwefel, Sphere)
